Speed "Fast Fall 2"
EDest has cooked up a new setting for speed, "Fast Fall 2". Works different than "Fast Fall" and bypasses much better.
Still working on diagonal jump scaffold. Maybe stop jumping for a couple seconds every once in a while? You should be able to go a good stack in a straight line when sprint jumping and easily twice as long diagonally, but the longer you go the more you're risking a kick or ban. This does not apply when not jumping or with the bypass disabled.
The way "Better Mode" works allows you to use it as an alternative to the non-jumping bypass with the benefit of being able to jump when needed.